Extract automation store row mappers and route helpers
